Consolidate in one io.py module.
More timing tests didn't show much impact of importing numpy or
not. All my evidence that "import numpy" was slow was anecdotal
and didn't hold up to testing. Maybe it *was* slower before 1.9,
I dunno.
