+++ rasterio/io.py
